Check for false in proc
In rails/rails@394b7be parameters set were changed to query strings to
behave more like a real browser. In a browser the parameters `false`
will never get sent as `false` and will always be `"false"`. Because
of this change the test for passing `false` into this `Proc` was
failing. We need to change it to check if the layout is not false rather
than relying on the value in the parameters.